feat(album): move items from a dialog, and remember the destination
Moving album items opened a route of its own. The selection that started it
belonged to the album tab, so reaching that page meant leaving the selection
behind, and there was no way back to it — the browser's back button returned
to an album with nothing selected.

Replace the page with a dialog raised from the album tab: full screen on a
phone, a centred dialog once there is room, the shell the image editor and
media picker already use. Escape and the backdrop close it.

The page had to carry the selection across a navigation, so the item ids went
through sessionStorage with a scopeKey query parameter and the navigation
state as fallbacks, and the page then refetched the project and the whole
album to turn those ids back into items. A dialog is handed the items it
already has, so all of that goes, along with the route and the page itself.

Its two-column layout does not fit a dialog, so the content is one column with
the dialog body as the only scroller: the preview shows twelve thumbnails and
collapses the rest into a count, and the candidate list is no longer capped,
since the search box is what narrows it. Nothing was dropped — the item count
and size, what travels with the move, the warning, the existing/new switch and
the new project name are all still there.

The destination is also remembered now, per project type. It defaulted to
whichever candidate sorted first, so moving a second batch to the same project
meant finding it again. Only a completed move is recorded, so browsing the
list and backing out does not change what the next one proposes, and a
destination created on the spot is recorded too. A remembered project that
sits below the fold is scrolled into view, since a selection nobody can see
reads as no selection at all.

Four translation keys the page owned are dropped from all six locales.
