+++ src/locales/en/project-viewer.json
@@ -307,11 +307,7 @@
       "moveSelected": "Move to Project",
       "moveShort": "Move",
       "title": "Move to Another Project",
-      "description": "Review what moves out of {{name}}, then choose where it lands.",
-      "back": "Back to album",
       "loadFailed": "Failed to load the move page",
-      "nothingSelected": "No items to move",
-      "nothingSelectedHint": "Select album items first, then choose Move to Project.",
       "itemsHeading_one": "{{count}} item selected",
       "itemsHeading_other": "{{count}} items selected",
       "includesItems_one": "{{count}} album item and its files",
