+++ src/locales/fr/project-viewer.json
@@ -307,11 +307,7 @@
       "moveSelected": "Déplacer vers un projet",
       "moveShort": "Déplacer",
       "title": "Déplacer vers un autre projet",
-      "description": "Vérifiez ce qui quitte {{name}}, puis choisissez la destination.",
-      "back": "Retour à l'album",
       "loadFailed": "Échec du chargement de la page de déplacement",
-      "nothingSelected": "Aucun élément à déplacer",
-      "nothingSelectedHint": "Sélectionnez d'abord des éléments de l'album, puis choisissez Déplacer vers un projet.",
       "itemsHeading_one": "{{count}} élément sélectionné",
       "itemsHeading_other": "{{count}} éléments sélectionnés",
       "includesItems_one": "{{count}} élément d'album et ses fichiers",
