+++ src/locales/ja/project-viewer.json
@@ -307,11 +307,7 @@
       "moveSelected": "プロジェクトへ移動",
       "moveShort": "移動",
       "title": "別のプロジェクトへ移動",
-      "description": "{{name}} から移動する内容を確認し、移動先を選択します。",
-      "back": "アルバムに戻る",
       "loadFailed": "移動ページの読み込みに失敗しました",
-      "nothingSelected": "移動する項目がありません",
-      "nothingSelectedHint": "先にアルバムの項目を選択してから「プロジェクトへ移動」を選んでください。",
       "itemsHeading_one": "{{count}} 件を選択中",
       "itemsHeading_other": "{{count}} 件を選択中",
       "includesItems_one": "アルバム項目 {{count}} 件とそのファイル",
