+++ src/locales/ko/project-viewer.json
@@ -307,11 +307,7 @@
       "moveSelected": "프로젝트로 이동",
       "moveShort": "이동",
       "title": "다른 프로젝트로 이동",
-      "description": "{{name}}에서 이동할 항목을 확인한 뒤 대상 프로젝트를 선택하세요.",
-      "back": "앨범으로 돌아가기",
       "loadFailed": "이동 페이지를 불러오지 못했습니다",
-      "nothingSelected": "이동할 항목이 없습니다",
-      "nothingSelectedHint": "앨범 항목을 먼저 선택한 뒤 '프로젝트로 이동'을 선택하세요.",
       "itemsHeading_one": "{{count}}개 선택됨",
       "itemsHeading_other": "{{count}}개 선택됨",
       "includesItems_one": "앨범 항목 {{count}}개와 해당 파일",
