+++ src/locales/zh-CN/project-viewer.json
@@ -307,11 +307,7 @@
       "moveSelected": "移动到项目",
       "moveShort": "移动",
       "title": "移动到其他项目",
-      "description": "确认将从 {{name}} 移出的内容，然后选择目标项目。",
-      "back": "返回相册",
       "loadFailed": "加载移动页面失败",
-      "nothingSelected": "没有可移动的项目",
-      "nothingSelectedHint": "请先选择相册条目，再选择“移动到项目”。",
       "itemsHeading_one": "已选择 {{count}} 个条目",
       "itemsHeading_other": "已选择 {{count}} 个条目",
       "includesItems_one": "{{count}} 个相册条目及其文件",
