+++ src/locales/zh-TW/project-viewer.json
@@ -307,11 +307,7 @@
       "moveSelected": "移動到專案",
       "moveShort": "移動",
       "title": "移動到其他專案",
-      "description": "確認將從 {{name}} 移出的內容，然後選擇目標專案。",
-      "back": "返回相簿",
       "loadFailed": "載入移動頁面失敗",
-      "nothingSelected": "沒有可移動的項目",
-      "nothingSelectedHint": "請先選擇相簿項目，再選擇「移動到專案」。",
       "itemsHeading_one": "已選擇 {{count}} 個項目",
       "itemsHeading_other": "已選擇 {{count}} 個項目",
       "includesItems_one": "{{count}} 個相簿項目及其檔案",
